move input types tests to own spec
